+++ tests/config/wildcard-example.rs
@@ -1,0 +1,5 @@
+group_imports = [
+    ["$std::*", "proc_macro::*"],
+    ["my_crate::*", "crate::*::xyz"],
+    ["$crate::*"],
+]