+++ tests/target/wildcard-group-import/config-example.rs
@@ -1,0 +1,12 @@
+// rustfmt-unstable: true
+// rustfmt-config: wildcard-example.rs
+use proc_macro::Span;
+use std::rc::Rc;
+
+use crate::abc::xyz;
+use my_crate::a::B;
+use my_crate::A;
+
+use self::X;
+use super::Y;
+use crate::Z;