+++ tests/target/wildcard-group-import/import.rs
@@ -1,0 +1,19 @@
+// rustfmt-unstable: true
+// rustfmt-config: wildcard-base.toml
+use a;
+use b::c;
+
+use crate::x::y;
+
+use core::sync;
+use std::sync;
+
+use a::ipsum::b;
+use c::d::ipsum::b;
+
+use xyz::abc::dez;
+
+use self::x;
+use crate::y;
+use b::c::d;
+use core::sync::A;